requirements.txt.txt
streamlit
google-generativeai
pillow
python-dotenv